finally end our long great national two-template elevation total nightmare
